## Summary

Upgrade go-cron to v0.6.0 and implement new scheduler capabilities:

- **Graceful Shutdown**: Add `StopWithTimeout()` and `StopAndWait()`
methods for controlled scheduler termination with job completion waiting
- **Job Naming**: Use `WithName()` for O(1) job lookup via
`cron.EntryByName()`
- **Job Tagging**: Add `AddJobWithTags()`, `RemoveJobsByTag()`,
`GetJobsByTag()` for logical job grouping and batch operations
- **ObservabilityHooks**: Integrate `JobStarted`, `JobCompleted`,
`JobScheduled` callbacks with MetricsRecorder for comprehensive job
monitoring
- **Fix DisableJob/EnableJob**: Handle go-cron's async removal with
exponential backoff polling in EnableJob
- **Update stress tests**: Pre-add jobs before starting scheduler to
avoid go-cron internal race conditions, add staggered worker starts

## Test Plan

- [x] All existing tests pass
- [x] Race detection passes for concurrent operations
- [x] New go-cron v0.6.0 features integrated and tested
- [x] DisableJob/EnableJob cycle works correctly
- [x] Graceful shutdown waits for running jobs

## Changes

- `core/scheduler.go`: New scheduler methods leveraging go-cron v0.6.0
- `core/scheduler_edge_cases_test.go`: Updated stress test to avoid race
conditions
- `core/retry.go`: Extended MetricsRecorder interface for
ObservabilityHooks
- `metrics/prometheus.go`: Added nolint directive for getGaugeValue
- `go.mod`, `go.sum`: go-cron v0.6.0 dependency

+++ core/common.go
@@ -31,8 +31,8 @@ type Job interface {
 	Running() int32
 	NotifyStart()
 	NotifyStop()
-	GetCronJobID() int
-	SetCronJobID(int)
+	GetCronJobID() uint64
+	SetCronJobID(uint64)
 	GetHistory() []*Execution
 	Hash() (string, error)
 }
